requirements.txt
Flask==3.0.2
paho-mqtt==1.6.1
